Set appropriate StockOrderQuantities when deleting transaction

+++ src/main/java/com/abelium/inatrace/db/entities/stockorder/Transaction.java
@@ -65,10 +65,10 @@ public class Transaction extends TimestampEntity {
 	private MeasureUnitType outputMeasureUnitType;
 	
 	@Column
-	private BigDecimal inputQuantity;
+	private Integer inputQuantity;
 
 	@Column
-	private BigDecimal outputQuantity;
+	private Integer outputQuantity;
 	
 	@Column
 	private BigDecimal pricePerUnit;
@@ -194,19 +194,19 @@ public class Transaction extends TimestampEntity {
 		this.outputMeasureUnitType = outputMeasureUnitType;
 	}
 
-	public BigDecimal getInputQuantity() {
+	public Integer getInputQuantity() {
 		return inputQuantity;
 	}
 
-	public void setInputQuantity(BigDecimal inputQuantity) {
+	public void setInputQuantity(Integer inputQuantity) {
 		this.inputQuantity = inputQuantity;
 	}
 
-	public BigDecimal getOutputQuantity() {
+	public Integer getOutputQuantity() {
 		return outputQuantity;
 	}
 
-	public void setOutputQuantity(BigDecimal outputQuantity) {
+	public void setOutputQuantity(Integer outputQuantity) {
 		this.outputQuantity = outputQuantity;
 	}
